+++ Frameworks/AVFoundation.framework/AVAssetDownloadSessionInternal.h
@@ -13,9 +13,9 @@
     struct OpaqueFigAsset { } * figAsset;
     struct OpaqueFigPlaybackItem { } * playbackItem;
     struct OpaqueFigPlayer { } * player;
-    int  priority;
+    long long  priority;
     NSObject<OS_dispatch_queue> * readWriteQueue;
-    int  status;
+    long long  status;
     AVWeakReference * weakReference;
 }
