+++ Frameworks/AVFoundation.framework/AVAssetReaderOutputMetadataAdaptor.h
@@ -10,7 +10,7 @@
 
 + (id)assetReaderOutputMetadataAdaptorWithAssetReaderTrackOutput:(id)arg1;
 
-- (long)addExtractionForOutput:(id)arg1 figAssetReader:(struct OpaqueFigAssetReader { }*)arg2 options:(id)arg3 withOutputExtactionID:(int*)arg4;
+- (int)addExtractionForOutput:(id)arg1 figAssetReader:(struct OpaqueFigAssetReader { }*)arg2 options:(id)arg3 withOutputExtactionID:(int*)arg4;
 - (id)assetReaderTrackOutput;
 - (void)dealloc;
 - (void)finalize;
