+++ Frameworks/AVFoundation.framework/AVCallbackContextRegistry.h
@@ -4,7 +4,7 @@
 
 @interface AVCallbackContextRegistry : NSObject {
     NSMutableDictionary * _contextsForTokens;
-    unsigned long  _currentToken;
+    unsigned long long  _currentToken;
     NSObject<OS_dispatch_queue> * _readWriteQueue;
 }
