+++ Frameworks/AVFoundation.framework/AVCaptureDeviceControlRequest.h
@@ -4,12 +4,12 @@
 
 @interface AVCaptureDeviceControlRequest : NSObject {
     id /* block */  _completionBlock;
-    long  _errorCode;
+    int  _errorCode;
     int  _requestID;
 }
 
 @property (nonatomic, readonly, copy) id /* block */ completionBlock;
-@property (nonatomic) long errorCode;
+@property (nonatomic) int errorCode;
 @property (nonatomic, readonly) int requestID;
 
 + (id)deviceControlRequestWithCompletionBlock:(id /* block */)arg1;
@@ -17,8 +17,8 @@
 - (id)_initWithCompletionBlock:(id /* block */)arg1;
 - (id /* block */)completionBlock;
 - (void)dealloc;
-- (long)errorCode;
+- (int)errorCode;
 - (int)requestID;
-- (void)setErrorCode:(long)arg1;
+- (void)setErrorCode:(int)arg1;
 
 @end
