+++ Frameworks/AVFoundation.framework/AVCaptureDeviceDiscoverySession.h
@@ -6,14 +6,14 @@
     NSArray * _deviceTypes;
     NSArray * _devices;
     NSString * _mediaType;
-    int * _position;
+    long long * _position;
 }
 
 @property (nonatomic, readonly) NSArray *devices;
 
-+ (id)discoverySessionWithDeviceTypes:(id)arg1 mediaType:(id)arg2 position:(int)arg3;
++ (id)discoverySessionWithDeviceTypes:(id)arg1 mediaType:(id)arg2 position:(long long)arg3;
 
-- (id)_initWithDeviceTypes:(id)arg1 mediaType:(id)arg2 position:(int)arg3;
+- (id)_initWithDeviceTypes:(id)arg1 mediaType:(id)arg2 position:(long long)arg3;
 - (void)dealloc;
 - (id)description;
 - (id)devices;
