+++ Frameworks/AVFoundation.framework/AVDataAsset.h
@@ -3,13 +3,13 @@
  */
 
 @interface AVDataAsset : AVAsset {
-    long  _allocateTracksOnceOnly;
+    long long  _allocateTracksOnceOnly;
     NSDictionary * _initializationOptions;
     AVAssetInspectorLoader * _loader;
     NSArray * _tracks;
 }
 
-+ (unsigned int)_dataLengthLimit;
++ (unsigned long long)_dataLengthLimit;
 + (id)_getFigAssetCreationOptionsFromDataAssetInitializationOptions:(id)arg1 figAssetCreationFlags:(unsigned long long*)arg2;
 
 - (id)_assetInspector;
@@ -19,7 +19,7 @@
 - (void)dealloc;
 - (id)initWithData:(id)arg1 contentType:(id)arg2;
 - (id)initWithData:(id)arg1 contentType:(id)arg2 options:(id)arg3;
-- (unsigned int)referenceRestrictions;
+- (unsigned long long)referenceRestrictions;
 - (id)tracks;
 
 @end
