+++ Frameworks/AVFoundation.framework/AVFigObjectInspector.h
@@ -7,8 +7,8 @@
 + (void)initialize;
 
 - (short)_SInt16ForProperty:(struct __CFString { }*)arg1;
-- (long)_SInt32ForProperty:(struct __CFString { }*)arg1;
-- (struct CGAffineTransform { float x1; float x2; float x3; float x4; float x5; float x6; })_affineTransformForProperty:(struct __CFString { }*)arg1;
+- (int)_SInt32ForProperty:(struct __CFString { }*)arg1;
+- (struct CGAffineTransform { double x1; double x2; double x3; double x4; double x5; double x6; })_affineTransformForProperty:(struct __CFString { }*)arg1;
 - (id)_arrayForProperty:(struct __CFString { }*)arg1;
 - (unsigned char)_booleanForProperty:(struct __CFString { }*)arg1;
 - (id)_dataForProperty:(struct __CFString { }*)arg1;
@@ -18,8 +18,8 @@
 - (long long)_longLongForProperty:(struct __CFString { }*)arg1;
 - (id)_nonNilArrayForProperty:(struct __CFString { }*)arg1;
 - (id)_nonNilDictionaryForProperty:(struct __CFString { }*)arg1;
-- (struct CGSize { float x1; float x2; })_sizeForProperty:(struct __CFString { }*)arg1;
-- (struct CGSize { float x1; float x2; })_sizeForProperty:(struct __CFString { }*)arg1 defaultValue:(struct CGSize { float x1; float x2; })arg2;
+- (struct CGSize { double x1; double x2; })_sizeForProperty:(struct __CFString { }*)arg1;
+- (struct CGSize { double x1; double x2; })_sizeForProperty:(struct __CFString { }*)arg1 defaultValue:(struct CGSize { double x1; double x2; })arg2;
 - (id)_stringForProperty:(struct __CFString { }*)arg1;
 - (struct { long long x1; int x2; unsigned int x3; long long x4; })_timeForProperty:(struct __CFString { }*)arg1;
 - (struct { long long x1; int x2; unsigned int x3; long long x4; })_timeForProperty:(struct __CFString { }*)arg1 defaultValue:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg2;
