+++ Frameworks/AVFoundation.framework/AVMediaSelectionOptionInternal.h
@@ -7,7 +7,7 @@
     NSString * _displayNameLocaleIdentifier;
     NSArray * _groupMediaCharacteristics;
     NSArray * _optionMediaCharacteristics;
-    long  _synthesizeMediaCharacteristicsOnce;
+    long long  _synthesizeMediaCharacteristicsOnce;
 }
 
 @end
