+++ Frameworks/AVFoundation.framework/AVUnreachableAssetInspectorLoader.h
@@ -3,14 +3,14 @@
  */
 
 @interface AVUnreachableAssetInspectorLoader : AVAssetInspectorLoader {
-    long  _figError;
+    int  _figError;
     NSDictionary * _figErrorUserInfo;
 }
 
 - (Class)_classForTrackInspectors;
 - (void)dealloc;
-- (id)initWithFigError:(long)arg1 userInfo:(id)arg2;
+- (id)initWithFigError:(int)arg1 userInfo:(id)arg2;
 - (void)loadValuesAsynchronouslyForKeys:(id)arg1 completionHandler:(id /* block */)arg2;
-- (int)statusOfValueForKey:(id)arg1 error:(id*)arg2;
+- (long long)statusOfValueForKey:(id)arg1 error:(id*)arg2;
 
 @end
