+++ Frameworks/AssetsLibrary.framework/ALAssetsGroupList.h
@@ -8,10 +8,10 @@
 
 @property (nonatomic, readonly, retain) <PLAssetContainerList> *containerList;
 
-- (void)_enumerateAssetsAtIndexes:(id)arg1 options:(unsigned int)arg2 usingBlock:(id /* block */)arg3;
+- (void)_enumerateAssetsAtIndexes:(id)arg1 options:(unsigned long long)arg2 usingBlock:(id /* block */)arg3;
 - (id)containerList;
 - (void)dealloc;
-- (id)initWithContainerList:(id)arg1 library:(id)arg2 type:(unsigned int)arg3;
+- (id)initWithContainerList:(id)arg1 library:(id)arg2 type:(unsigned long long)arg3;
 - (id)valueForProperty:(id)arg1;
 
 @end
