+++ Frameworks/CFNetwork.framework/NSURLAuthenticationChallengeInternal.h
@@ -5,8 +5,8 @@
 @interface NSURLAuthenticationChallengeInternal : NSObject {
     NSError * error;
     NSURLResponse * failureResponse;
-    int  preferredProtSpaceIndex;
-    int  previousFailureCount;
+    long long  preferredProtSpaceIndex;
+    long long  previousFailureCount;
     NSURLCredential * proposedCredential;
     NSArray * protectionSpacesForChallenge;
     <NSURLAuthenticationChallengeSender> * sender;
@@ -14,6 +14,6 @@
 }
 
 - (void)dealloc;
-- (id)initWithProtectionSpace:(id)arg1 proposedCredential:(id)arg2 previousFailureCount:(int)arg3 failureResponse:(id)arg4 error:(id)arg5 sender:(id)arg6;
+- (id)initWithProtectionSpace:(id)arg1 proposedCredential:(id)arg2 previousFailureCount:(long long)arg3 failureResponse:(id)arg4 error:(id)arg5 sender:(id)arg6;
 
 @end
