+++ Frameworks/Contacts.framework/CNFamilyNameFirstNameOrder.h
@@ -4,7 +4,7 @@
 
 @interface CNFamilyNameFirstNameOrder : CNContactNameOrderImpl
 
-- (int)nameOrder;
+- (long long)nameOrder;
 - (id)nameProperties;
 - (id)phoneticNameProperties;
 - (id)sortingNameProperties;
