+++ Frameworks/ContactsUI.framework/CNAvatarCardActionsTableView.h
@@ -4,6 +4,6 @@
 
 @interface CNAvatarCardActionsTableView : UITableView
 
-- (struct CGSize { float x1; float x2; })intrinsicContentSize;
+- (struct CGSize { double x1; double x2; })intrinsicContentSize;
 
 @end
