+++ Frameworks/ContactsUI.framework/CNContactCreateNewContactAction.h
@@ -10,7 +10,7 @@
 @property (nonatomic, retain) CNContact *createdContact;
 @property (readonly, copy) NSString *debugDescription;
 @property (readonly, copy) NSString *description;
-@property (readonly) unsigned int hash;
+@property (readonly) unsigned long long hash;
 @property (nonatomic) UIViewController *presentingViewController;
 @property (readonly) Class superclass;
