+++ Frameworks/ContactsUI.framework/CNCustomActionSheetPresentation.h
@@ -4,6 +4,6 @@
 
 @interface CNCustomActionSheetPresentation : NSObject
 
-+ (id)prepareViewControllerForPresentation:(id)arg1 sourceView:(id)arg2 sourceRect:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg3 permittedArrowDirection:(unsigned int)arg4 dismissHandler:(id /* block */)arg5;
++ (id)prepareViewControllerForPresentation:(id)arg1 sourceView:(id)arg2 sourceRect:(struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })arg3 permittedArrowDirection:(unsigned long long)arg4 dismissHandler:(id /* block */)arg5;
 
 @end
