+++ Frameworks/ContactsUI.framework/_CNCustomActionSheetPresentation.h
@@ -4,15 +4,15 @@
 
 @interface _CNCustomActionSheetPresentation : NSObject <CNCustomPresentation, UIActionSheetPresentationControllerDelegate, UIViewControllerTransitioningDelegate> {
     id /* block */  _dismissHandler;
-    unsigned int  _permittedArrowDirections;
+    unsigned long long  _permittedArrowDirections;
     struct CGRect { 
         struct CGPoint { 
-            float x; 
-            float y; 
+            double x; 
+            double y; 
         } origin; 
         struct CGSize { 
-            float width; 
-            float height; 
+            double width; 
+            double height; 
         } size; 
     }  _sourceRect;
     UIView * _sourceView;
@@ -22,9 +22,9 @@
 @property (readonly, copy) NSString *debugDescription;
 @property (readonly, copy) NSString *description;
 @property (nonatomic, copy) id /* block */ dismissHandler;
-@property (readonly) unsigned int hash;
-@property (nonatomic) unsigned int permittedArrowDirections;
-@property (nonatomic) struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; } sourceRect;
+@property (readonly) unsigned long long hash;
+@property (nonatomic) unsigned long long permittedArrowDirections;
+@property (nonatomic) struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; } sourceRect;
 @property (nonatomic, retain) UIView *sourceView;
 @property (readonly) Class superclass;
 @property (nonatomic, retain) UIViewController *viewController;
@@ -32,16 +32,16 @@
 - (void).cxx_destruct;
 - (void)actionSheetPresentationControllerDidDismissActionSheet:(id)arg1;
 - (id /* block */)dismissHandler;
-- (unsigned int)permittedArrowDirections;
+- (unsigned long long)permittedArrowDirections;
 - (void)prepareForPresentationOfController:(id)arg1;
 - (id)presentationControllerForPresentedViewController:(id)arg1 presentingViewController:(id)arg2 sourceViewController:(id)arg3;
 - (id)presentedViewController;
 - (void)setDismissHandler:(id /* block */)arg1;
-- (void)setPermittedArrowDirections:(unsigned int)arg1;
-- (void)setSourceRect:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
+- (void)setPermittedArrowDirections:(unsigned long long)arg1;
+- (void)setSourceRect:(struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })arg1;
 - (void)setSourceView:(id)arg1;
 - (void)setViewController:(id)arg1;
-- (struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })sourceRect;
+- (struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })sourceRect;
 - (id)sourceView;
 - (id)viewController;
