+++ Frameworks/CoreBluetooth.framework/CBScalablePipe.h
@@ -10,9 +10,9 @@
     NSOutputStream * _output;
     CBPeer * _peer;
     CBScalablePipeManager * _pipeManager;
-    int  _priority;
+    long long  _priority;
     int  _socket;
-    int  _type;
+    long long  _type;
 }
 
 @property (nonatomic, readonly) CBCentral *central;
@@ -21,8 +21,8 @@
 @property (nonatomic, readonly) NSString *name;
 @property (nonatomic, readonly) NSOutputStream *output;
 @property (nonatomic, readonly) CBPeer *peer;
-@property (nonatomic, readonly) int priority;
-@property (nonatomic, readonly) int type;
+@property (nonatomic, readonly) long long priority;
+@property (nonatomic, readonly) long long type;
 
 - (void).cxx_destruct;
 - (id)central;
@@ -34,8 +34,8 @@
 - (id)name;
 - (id)output;
 - (id)peer;
-- (int)priority;
+- (long long)priority;
 - (void)setOrphan;
-- (int)type;
+- (long long)type;
 
 @end
